Change name of the module (and so also the related packages) to flux-movement-plugin. The registered name (Exchange tab) will be "FLUX Movement Plugin"
