+++ module/pom.xml
@@ -8,7 +8,7 @@
 
     <parent>
         <groupId>eu.europa.ec.fisheries.uvms.plugins</groupId>
-        <artifactId>movement-plugin</artifactId>
+        <artifactId>flux-movement-plugin</artifactId>
         <version>3.0.13-SNAPSHOT</version>
     </parent>
