--- service/src/main/java/eu/europa/ec/fisheries/uvms/plugins/movement/exception/MappingException.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package eu.europa.ec.fisheries.uvms.plugins.movement.exception;
-
-/**
- *
- * @author jojoha
- */
-public class MappingException extends Exception {
-
-    public MappingException(String message) {
-        super(message);
-    }
-
-}
