--- service/src/main/java/eu/europa/ec/fisheries/uvms/plugins/movement/producer/PluginToEventBusTopicProducer.java
@@ -1,26 +1,0 @@
-/*
-Developed by the European Commission - Directorate General for Maritime Affairs and Fisheries @ European Union, 2015-2016.
-
-This file is part of the Integrated Fisheries Data Management (IFDM) Suite. The IFDM Suite is free software: you can redistribute it
-and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of
-the License, or any later version. The IFDM Suite is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-details. You should have received a copy of the GNU General Public License along with the IFDM Suite. If not, see <http://www.gnu.org/licenses/>.
-
-*/
-package eu.europa.ec.fisheries.uvms.plugins.movement.producer;
-
-import eu.europa.ec.fisheries.uvms.commons.message.api.MessageConstants;
-import eu.europa.ec.fisheries.uvms.commons.message.impl.AbstractTopicProducer;
-import javax.ejb.LocalBean;
-import javax.ejb.Stateless;
-
-@Stateless
-@LocalBean
-public class PluginToEventBusTopicProducer extends AbstractTopicProducer {
-
-    @Override
-    protected String getDestinationName() {
-        return MessageConstants.EVENT_BUS_TOPIC;
-    }
-}
