+++ service/src/main/resources/flux.properties
@@ -2,12 +2,12 @@
 #Wed Oct 28 19:55:28 CET 2015
 swagger.version=1.3.1
 sl4j.version=1.7.7
-application.logfile=/app/logs/movement-plugin/movement-plugin.log
-application.error.logfile=/app/logs/movement-plugin/movement-plugin-error.log
+application.logfile=/app/logs/movement-plugin/flux-movement-plugin.log
+application.error.logfile=/app/logs/movement-plugin/flux-movement-plugin-error.log
 junit.version=4.10
 application.responseTopicName=PLUGIN_RESPONSE
-application.name=movement-plugin
-application.groupid=eu.europa.ec.fisheries.uvms.plugins.movement
+application.name=FLUX Movement Plugin
+application.groupid=eu.europa.ec.fisheries.uvms.plugins.flux.movement
 maven.compiler.target=1.7
 exchange.model.version=1.0.2-SNAPSHOT
 logback.core.version=1.1.2
